Moving TranslationTrait to Translation namespace, Use PHPDoc from TranslationInterface
